Create receiveAttack function and test
Still not accounting for or testing what happens if the player guesses a
location not on the board.

Also, not accounting for or testing if the Ship doesn't make sense (too
long or short for the start and end coordinates, or start and end
coordinates could not be either vertical or horizontal ships).